requirements.txt
streamlit
torch
torchvision
pennylane
numpy
pillow
